Added node support to module wrapper, enabled support for anvil --mocha
Removed the node/ and browser/ folders. The lib directory now includes
two files: riveter.js (The wrapper) and riveter.base.js (The code).

Added support for `anvil --mocha` to run the tests in addition to the
existing browser version.

Added mocha to the dependency list, and a script `npm test` to run
the tests through mocha.